Make diagram type painter generic
Now the diagram package no longer depends on UML.
