+++ gaphor/SysML/__init__.py
@@ -1,8 +1,8 @@
 # ruff: noqa: F401
 
-import gaphor.SysML.diagramlabel
 import gaphor.SysML.drop
 import gaphor.SysML.modelinglanguage
+import gaphor.SysML.painter
 import gaphor.SysML.sysmlfmt
 
 __modeling_language__ = "SysML"
