+++ gaphor/SysML/tests/test_diagramlabel.py
@@ -1,31 +1,39 @@
-from gaphor.diagram.diagramlabel import diagram_label
+import pytest
+
+from gaphor.SysML.painter import SysMLDiagramTypePainter
 from gaphor.SysML.sysml import Block, ConstraintBlock, Requirement, SysMLDiagram
 from gaphor.UML.uml import Activity, Interaction, Package, Profile, StateMachine
 
 
-def test_diagram_label():
+@pytest.mark.parametrize(
+    "type,label",
+    [
+        [Activity, "[activity] ElemName [DiagName]"],
+        [Block, "[block] ElemName [DiagName]"],
+        [ConstraintBlock, "[constraint block] ElemName [DiagName]"],
+        [Interaction, "[interaction] ElemName [DiagName]"],
+        [Package, "[package] ElemName [DiagName]"],
+        [Profile, "[profile] ElemName [DiagName]"],
+        [Requirement, "[requirement] ElemName [DiagName]"],
+        [StateMachine, "[state machine] ElemName [DiagName]"],
+    ],
+)
+def test_diagram_label(type, label):
     def _diagram(elementType):
         element = elementType()
-        element.name = "Efghi"
+        element.name = "ElemName"
 
         diagram = SysMLDiagram()
-        diagram.name = "Abcd"
+        diagram.name = "DiagName"
         diagram.element = element
 
         return diagram
 
-    assert diagram_label(_diagram(Activity)) == "[activity] Efghi [Abcd]"
-    assert diagram_label(_diagram(Block)) == "[block] Efghi [Abcd]"
-    assert diagram_label(_diagram(ConstraintBlock)) == "[constraint block] Efghi [Abcd]"
-    assert diagram_label(_diagram(Interaction)) == "[interaction] Efghi [Abcd]"
-    assert diagram_label(_diagram(Package)) == "[package] Efghi [Abcd]"
-    assert diagram_label(_diagram(Profile)) == "[profile] Efghi [Abcd]"
-    assert diagram_label(_diagram(Requirement)) == "[requirement] Efghi [Abcd]"
-    assert diagram_label(_diagram(StateMachine)) == "[state machine] Efghi [Abcd]"
+    assert SysMLDiagramTypePainter(_diagram(type)).label() == label
 
 
 def test_uncontained_diagram_label():
     diagram = SysMLDiagram()
-    diagram.name = "Abcd"
+    diagram.name = "DiagName"
 
-    assert diagram_label(diagram) == "Abcd"
+    assert SysMLDiagramTypePainter(diagram).label() == "DiagName"
