--- gaphor/diagram/diagramlabel.py
@@ -1,6 +1,0 @@
-from functools import singledispatch
-
-
-@singledispatch
-def diagram_label(diagram):
-    return diagram.name
