+++ gaphor/diagram/painter.py
@@ -8,14 +8,12 @@ and handles).
 
 from __future__ import annotations
 
-from cairo import LINE_JOIN_ROUND
-from gaphas.geometry import Rectangle
+from functools import singledispatch
 
-from gaphor.core.modeling.diagram import DrawContext, StyledDiagram, StyledItem
-from gaphor.diagram.diagramlabel import diagram_label
+from cairo import LINE_JOIN_ROUND
+
+from gaphor.core.modeling.diagram import Diagram, DrawContext, StyledItem
 from gaphor.diagram.selection import Selection
-from gaphor.diagram.shapes import Box, CssNode, Orientation, Text, cairo_state, stroke
-from gaphor.UML.recipes import get_applied_stereotypes, stereotypes_str
 
 
 class ItemPainter:
@@ -58,57 +56,16 @@ class ItemPainter:
             self.paint_item(item, cr)
 
 
+@singledispatch
 class DiagramTypePainter:
-    def __init__(self, diagram):
-        self.diagram = diagram
-        self._pentagon = CssNode(
-            "pentagon",
-            diagram,
-            Box(
-                CssNode("diagramtype", None, Text(text=lambda: diagram.diagramType)),
-                CssNode(
-                    "stereotypes", None, Text(text=lambda: stereotypes_str(diagram))
-                ),
-                CssNode("name", None, Text(text=lambda: diagram_label(diagram))),
-                orientation=Orientation.HORIZONTAL,
-                draw=draw_pentagon,
-            ),
-        )
+    """Diagram painter.
 
-    def paint(self, _items, cr):
-        diagram = self.diagram
-        if not diagram.diagramType and not any(get_applied_stereotypes(diagram)):
-            return
+    Allows for drawing diagram specific admonitions,
+    such as is common for UML and SysML diagrams.
+    """
 
-        style = diagram.style(StyledDiagram(diagram))
+    def __init__(self, _diagram: Diagram):
+        pass
 
-        context = DrawContext(
-            cairo=cr,
-            style=style,
-            selected=False,
-            focused=False,
-            hovered=False,
-            dropzone=False,
-        )
-
-        with cairo_state(cr):
-            cr.identity_matrix()
-
-            size = self._pentagon.size(context)  # type: ignore[arg-type]
-            bb = Rectangle(0, 0, *size)
-            self._pentagon.draw(context, bb)
-
-
-def draw_pentagon(box, context, bounding_box):
-    cr = context.cairo
-    w = bounding_box.width
-    h = bounding_box.height
-    line_width = context.style.get("line-width", 2)
-    with cairo_state(cr):
-        h2 = h / 2.0
-        cr.move_to(-line_width, -line_width)
-        cr.line_to(-line_width, h)
-        cr.line_to(w - 4, h)
-        cr.line_to(w, h2)
-        cr.line_to(w, -line_width)
-        stroke(context, fill=True)
+    def paint(self, _items, _cr):
+        pass
