+++ tests/test_architecture.py
@@ -47,19 +47,12 @@ def test_core_packages():
 
 
 def test_diagram_package():
-    # NB1. gaphor.diagram.tools.dropzone includes gaphor.UML.recipes,
-    # so it can assign the right owner package to a newly created element.
-    # NB2. The image property page requires a file dialog and some error
-    # handling. Unfortunately this is UI code, so we need it from there.
     (
         archrule("Diagrams are part of the core")
         .match("gaphor.diagram*")
         .exclude("*.tests.*")
         .may_import(*GAPHOR_CORE)
         .may_import("gaphor.diagram*")
-        .may_import("gaphor.UML.compartments")
-        .may_import("gaphor.UML.recipes")
-        .may_import("gaphor.UML.uml")
         .should_not_import("gaphor*")
         .check(gaphor, skip_type_checking=True)
     )
